refactor(patina): split linter/engine into cohesive modules
